requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=0.24.0
requests>=2.26.0
pyyaml>=5.4.0
joblib
seaborn
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
